Timeline_flask_project/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-JWT-Extended==4.1.0
Flask-Cors==3.0.10
bcrypt==3.2.0
PyJWT==2.1.0
Werkzeug==2.0.1
opencv-python==4.5.3.56
ultralytics==8.0.20
pandas==1.3.3
scikit-learn==0.24.2
